simplify reasons
